Fix classification report calculation error
Previously, the calculation of precision, recall, and F1 score could all
result in a division by zero error if a class existed but no predictions
were made for that class. That commonly happened in the case where a
model only ever predicted the majority gesture.

The sklearn method `classification_report` takes in an optional
parameter `zero_division` which was previously set to `np.nan`. This
meant that any zero division errors for a class would result in a
precision/recall/f1-score for that class of `np.nan`. However, when
calculating the averages over all classes, `np.nan` values are omitted,
so if there are 50 `nan` values and one `0.99` value, the "average" will
be 0.99. Not great.

This fix sets the `zero_division` parameter to zero, reflecting that if
a model never predicts a class, then it fails in that class.